requirements.txt
torchaudio
torch
librosa
pytorch-lightning
hydra-core
tensorboardX
pandas
transformers[torch]
datasets
jiwer
orjson